Try to detect country automatically.
This is useful to ensure the wifi network is correctly working as expected, as
the valid bands vary per country.

There's a challenge as this effectively requires a network request, so network
must be setup to setup network. Try to do it on the host first via 'flash', as
the host likely has network connectivity, but also include auto-detection in
'setup.sh'.

Currently use ipinfo.io as it works great, but could be changed if necessary.